+++ shared/app/src/jvmMain/kotlin/com/yellowtrack/platform/core/di/PlatformModule.jvm.kt
@@ -1,5 +1,7 @@
 package com.yellowtrack.platform.core.di
 
+import com.yellowtrack.platform.core.common.storage.JvmVolumeInspector
+import com.yellowtrack.platform.core.common.storage.VolumeInspector
 import com.yellowtrack.platform.core.database.DatabaseDriverFactory
 import com.yellowtrack.platform.core.database.JvmDatabaseDriverFactory
 import com.yellowtrack.platform.core.export.DocumentSink
@@ -11,4 +13,5 @@ actual fun platformModule(): Module =
     module {
         single<DatabaseDriverFactory> { JvmDatabaseDriverFactory() }
         single<DocumentSink> { JvmDocumentSink() }
+        single<VolumeInspector> { JvmVolumeInspector() }
     }