+++ shared/app/src/wasmJsMain/kotlin/com/yellowtrack/platform/core/di/PlatformModule.wasmJs.kt
@@ -1,5 +1,7 @@
 package com.yellowtrack.platform.core.di
 
+import com.yellowtrack.platform.core.common.storage.VolumeInspector
+import com.yellowtrack.platform.core.common.storage.WebVolumeInspector
 import com.yellowtrack.platform.core.database.DatabaseDriverFactory
 import com.yellowtrack.platform.core.database.WebDatabaseDriverFactory
 import com.yellowtrack.platform.core.export.DocumentSink
@@ -12,4 +14,5 @@ actual fun platformModule(): Module =
         // Requires the SQLite worker script to be served by the web application.
         single<DatabaseDriverFactory> { WebDatabaseDriverFactory() }
         single<DocumentSink> { WebDocumentSink() }
+        single<VolumeInspector> { WebVolumeInspector() }
     }